refactor: finalize Search router and delete legacy SearchList shell
